separate out script furher for easier maintenance
